+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/BranchDto.java
@@ -35,15 +35,13 @@ import org.hibernate.validator.constraints.Length;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Pattern;
 
+import static sonia.scm.repository.Branch.VALID_BRANCH_NAMES;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @SuppressWarnings("java:S2160") // we do not need this for dto
 public class BranchDto extends HalRepresentation {
-
-  private static final String VALID_CHARACTERS_AT_START_AND_END = "\\w-,;\\]{}@&+=$#`|<>";
-  private static final String VALID_CHARACTERS = VALID_CHARACTERS_AT_START_AND_END + "/.";
-  static final String VALID_BRANCH_NAMES = "[" + VALID_CHARACTERS_AT_START_AND_END + "]([" + VALID_CHARACTERS + "]*[" + VALID_CHARACTERS_AT_START_AND_END + "])?";
 
   @NotEmpty
   @Length(min = 1, max = 100)